fixing main
