+++ src/main/java/WrightFlightManager/GUI/WelcomeScreen.java
@@ -1,0 +1,18 @@
+package WrightFlightManager.GUI;
+
+import javax.swing.*;
+
+public class WelcomeScreen extends JFrame {
+    public WelcomeScreen() {
+        setTitle("Welcome");
+        setSize(600, 600);
+        setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+        setLocationRelativeTo(null);
+        setResizable(false);
+        JLabel welcomeLabel = new JLabel("Welcome to Wright Flight Manager App");
+        welcomeLabel.setHorizontalAlignment(SwingConstants.CENTER);
+        add(welcomeLabel);
+
+        setVisible(true);
+    }
+}
