+++ src/main/java/WrightFlightManager/Main.java
@@ -1,5 +1,8 @@
 package WrightFlightManager;
+import WrightFlightManager.GUI.WelcomeScreen;
+
 import javax.swing.*;
+
 
 public class Main {
     public static void main(String[] args) {
@@ -10,17 +13,3 @@ public class Main {
         });
     }
 }
-
-public class WelcomeScreen extends JFrame {
-    public WelcomeScreen() {
-        setTitle("Welcome");
-        setSize(600, 600);
-        setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-        setLocationRelativeTo(null);
-        setResizable(false);
-        JLabel welcomeLabel = new JLabel("Welcome to Wright Flight Manager App");
-        welcomeLabel.setHorizontalAlignment(SwingConstants.CENTER);
-        add(welcomeLabel);
-
-        setVisible(true);
-    }
